+++ Changelog.md
@@ -2,6 +2,8 @@ v0.0.5
 ======
 - Asynchronous writes to disk.
 - Asynchronous memory storage.
+- PageManager: drop loaded chunks (on get_open_chunks) and dec chunks_in_cur_page() counter.
+- PageManager: page fragmentation support.
 
 v0.0.4
 ======